requirements.txt
fastapi
uvicorn
motor
python-dotenv
google-generativeai
httpx
Pillow
python-multipart